@@ -25,7 +25,7 @@
     "@nextcloud/initial-state": "^2.2.0",
     "@nextcloud/router": "^3.0.1",
     "@nextcloud/vue": "^9.5.0",
-    "vue": "^2.7.16",
+    "vue": "^3.5.27",
     "vue-material-design-icons": "^5.3.1"
   },
   "browserslist": [
@@ -44,7 +44,7 @@
     "@nextcloud/eslint-config": "^8.4.2",
     "@nextcloud/prettier-config": "^1.1.0",
     "@nextcloud/stylelint-config": "^3.0.1",
-    "@nextcloud/vite-config": "^1.6.0",
+    "@nextcloud/vite-config": "^2.5.2",
     "baseline-browser-mapping": "^2.9.19",
     "eslint-config-prettier": "^10.0.2",
     "prettier": "^3.5.3",
