@@ -3,12 +3,9 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import Vue from 'vue'
+import { createApp } from 'vue'
 import Settings from './components/UserSettings/Settings.vue'
 
-Vue.mixin({ methods: { t, n } })
-
-export default new Vue({
-	el: '#config-snowflakestheme',
-	render: (h) => h(Settings),
-})
+const app = createApp(Settings)
+app.mixin({ methods: { t, n } })
+app.mount('#config-snowflakestheme')
